+++ hal/posix/usb_devkit/libtropic_port_posix_usb_devkit.c
@@ -1,16 +1,16 @@
 /**
- * @file libtropic_port_posix_usb_dongle.c
+ * @file libtropic_port_posix_usb_devkit.c
  * @copyright Copyright (c) 2020-2026 Tropic Square s.r.o.
- * @brief Port for communication with USB UART Dongle (TS1302).
+ * @brief Port for communication with TROPIC01 USB DevKit.
  *
- * The TS1302 dongle uses a special protocol to translate UART communication to SPI. This port
- * implements the protocol. More info about the dongle in GitHub repo:
- * https://github.com/tropicsquare/ts13-usb-dev-kit-fw
+ * The TROPIC01 USB DevKit uses a special protocol to translate UART communication to SPI. This port
+ * implements the protocol. More info about the USB DevKit here:
+ * https://github.com/tropicsquare/devboards/tree/main/TROPIC01_USB_Devkit
  *
  * @license For the license see LICENSE.md in the root directory of this source tree.
  */
 
-#include "libtropic_port_posix_usb_dongle.h"
+#include "libtropic_port_posix_usb_devkit.h"
 
 #include <errno.h>
 #include <fcntl.h>
@@ -32,7 +32,7 @@
 #include "libtropic_port.h"
 
 #if LT_USE_INT_PIN
-#error "Interrupt PIN not supported in the USB dongle port!"
+#error "Interrupt PIN not supported in the USB DevKit port!"
 #endif
 
 // getentropy() has a limit of random bytes it can generate in one call. The POSIX.1-2024 standard
@@ -137,7 +137,7 @@ static bool read_port(int fd, uint8_t *buffer, size_t size)
 
 lt_ret_t lt_port_init(lt_l2_state_t *s2)
 {
-    lt_dev_posix_usb_dongle_t *device = (lt_dev_posix_usb_dongle_t *)s2->device;
+    lt_dev_posix_usb_devkit_t *device = (lt_dev_posix_usb_devkit_t *)s2->device;
 
     // Initialize the serial port.
     device->fd = open(device->dev_path, O_RDWR | O_NOCTTY);
@@ -210,7 +210,7 @@ lt_ret_t lt_port_init(lt_l2_state_t *s2)
 
 lt_ret_t lt_port_deinit(lt_l2_state_t *s2)
 {
-    lt_dev_posix_usb_dongle_t *device = (lt_dev_posix_usb_dongle_t *)s2->device;
+    lt_dev_posix_usb_devkit_t *device = (lt_dev_posix_usb_devkit_t *)s2->device;
 
     if (close(device->fd)) {
         return LT_FAIL;
@@ -264,7 +264,7 @@ lt_ret_t lt_port_spi_csn_low(lt_l2_state_t *s2)
 
 lt_ret_t lt_port_spi_csn_high(lt_l2_state_t *s2)
 {
-    lt_dev_posix_usb_dongle_t *device = (lt_dev_posix_usb_dongle_t *)s2->device;
+    lt_dev_posix_usb_devkit_t *device = (lt_dev_posix_usb_devkit_t *)s2->device;
 
     uint8_t cs_high[] = "CS=0\n";  // Yes, CS=0 really means that CSN is low
     if (!write_port(device->fd, cs_high, 5)) {
@@ -290,7 +290,7 @@ lt_ret_t lt_port_spi_transfer(lt_l2_state_t *s2, uint8_t offset, uint16_t tx_dat
 {
     LT_UNUSED(timeout_ms);
 
-    lt_dev_posix_usb_dongle_t *device = (lt_dev_posix_usb_dongle_t *)s2->device;
+    lt_dev_posix_usb_devkit_t *device = (lt_dev_posix_usb_devkit_t *)s2->device;
 
     if (offset + tx_data_length > TR01_L1_LEN_MAX) {
         return LT_L1_DATA_LEN_ERROR;
@@ -298,13 +298,13 @@ lt_ret_t lt_port_spi_transfer(lt_l2_state_t *s2, uint8_t offset, uint16_t tx_dat
 
     // Bytes from handle which are about to be sent are encoded as chars and stored to
     // buffered_chars.
-    uint8_t buffered_chars[LT_USB_DONGLE_SPI_TRANSFER_BUFF_SIZE_MAX] = {0};
+    uint8_t buffered_chars[LT_USB_DEVKIT_SPI_TRANSFER_BUFF_SIZE_MAX] = {0};
     for (int i = 0; i < tx_data_length; i++) {
         sprintf((char *)(buffered_chars + i * 2), "%02" PRIX8, s2->buff[i + offset]);
     }
 
-    // Control characters to keep CS LOW (they are expected by USB dongle, see the top of this file
-    // for more information).
+    // Control characters to keep CS LOW (they are expected by USB DevKit, see the top of this file for
+    // more information).
     buffered_chars[tx_data_length * 2] = 'x';
     buffered_chars[tx_data_length * 2 + 1] = '\n';
 
@@ -313,7 +313,7 @@ lt_ret_t lt_port_spi_transfer(lt_l2_state_t *s2, uint8_t offset, uint16_t tx_dat
         return LT_L1_SPI_ERROR;
     }
 
-    lt_port_delay(s2, LT_USB_DONGLE_READ_WRITE_DELAY);
+    lt_port_delay(s2, LT_USB_DEVKIT_READ_WRITE_DELAY);
 
     if (!read_port(device->fd, buffered_chars, (2 * tx_data_length) + 2)) {
         LT_LOG_ERROR("Failed to read SPI payload.");